Avoid use of spsc::read_available
Legacy boost didn't supply a read_available method, only a "non-precise"
empty() method. (This is despite the fact that the legacy version was
perfectly capable of strictly determining this, and in fact does so in
its pop methods).

Given PipeSubscriber is already protected by its state-model this
releaxed ordering is fine, and is technically more performant.