Aggregated Pipe - Initial Late Subscriber support
New subscribers to the aggregated pipe will now an initial set of data,
representing what is currently in scope - as a series of NEW updates.

NOTE: Whilst the underlying PipePub / PipeSub mechanism is safe, the
AggPipePub is currently not thread safe due to concurrent access to the
data map. A follow up change is required to tidy this up