AggPubPipe - support supressing update diffing.
By default the AggPub attempts to diff each update, in order to
determine if it should actually publish an Update. This requires the
Message object to implement IsAggEqual.

It is now possible to instantiate the AggPub in a mode which disables
the checking. In this mode, it is not required for the Message object to
implement the IsAggEqual method.